Move PbrBundle onto TerrainBundle
