+++ emergence_lib/src/graphics/mod.rs
@@ -9,7 +9,6 @@ use self::lighting::LightingPlugin;
 mod lighting;
 mod selection;
 mod structures;
-mod terrain;
 mod units;
 
 /// Adds all logic required to render the game.
@@ -22,7 +21,6 @@ impl Plugin for GraphicsPlugin {
         app.add_plugin(LightingPlugin)
             .add_system_set(
                 SystemSet::on_update(AssetState::Ready)
-                    .with_system(terrain::populate_terrain)
                     .with_system(units::populate_units)
                     .with_system(units::display_held_item)
                     // We need to avoid attempting to insert bundles into entities that no longer exist
