+++ emergence_lib/src/simulation/generation.rs
@@ -1,4 +1,5 @@
 //! Generating starting terrain and organisms
+use crate::asset_management::terrain::TerrainHandles;
 use crate::enum_iter::IterableEnum;
 use crate::items::ItemId;
 use crate::organisms::energy::{Energy, EnergyPool};
@@ -142,6 +143,7 @@ const SEED: f32 = 2378.0;
 pub(crate) fn generate_terrain(
     mut commands: Commands,
     config: Res<GenerationConfig>,
+    handles: Res<TerrainHandles>,
     mut map_geometry: ResMut<MapGeometry>,
 ) {
     info!("Generating terrain...");
@@ -167,13 +169,20 @@ pub(crate) fn generate_terrain(
                 // Height is stepped, and should always be a multiple of 1.0
                 .round();
 
+        // Store the height, so it can be used below
+        map_geometry.height_index.insert(tile_pos, hex_height);
+
         // Spawn the terrain entity
         let terrain_entity = commands
-            .spawn(TerrainBundle::new(terrain_type, tile_pos))
+            .spawn(TerrainBundle::new(
+                terrain_type,
+                tile_pos,
+                &handles,
+                &map_geometry,
+            ))
             .id();
 
-        // Update the index
-        map_geometry.height_index.insert(tile_pos, hex_height);
+        // Update the index of what terrain is where
         map_geometry.terrain_index.insert(tile_pos, terrain_entity);
     }
 }
