+++ emergence_lib/src/terrain/mod.rs
@@ -5,8 +5,9 @@ use bevy_mod_raycast::RaycastMesh;
 
 use crate as emergence_lib;
 
+use crate::asset_management::terrain::TerrainHandles;
 use crate::player_interaction::zoning::Zoning;
-use crate::simulation::geometry::TilePos;
+use crate::simulation::geometry::{MapGeometry, TilePos};
 use bevy::ecs::component::Component;
 use derive_more::Display;
 
@@ -52,16 +53,41 @@ pub(crate) struct TerrainBundle {
     raycast_mesh: RaycastMesh<Terrain>,
     /// The structure that should be built here.
     zoning: Zoning,
+    /// The mesh and material used
+    pbr_bundle: PbrBundle,
 }
 
 impl TerrainBundle {
     /// Creates a new Terrain entity.
-    pub(crate) fn new(terrain_type: Terrain, tile_pos: TilePos) -> Self {
+    pub(crate) fn new(
+        terrain_type: Terrain,
+        tile_pos: TilePos,
+        handles: &TerrainHandles,
+        map_geometry: &MapGeometry,
+    ) -> Self {
+        let world_pos = tile_pos.into_world_pos(&map_geometry);
+
+        let pbr_bundle = PbrBundle {
+            mesh: handles.mesh.clone_weak(),
+            material: handles
+                .terrain_materials
+                .get(&terrain_type)
+                .unwrap()
+                .clone_weak(),
+            transform: Transform::from_xyz(world_pos.x, 0.0, world_pos.z).with_scale(Vec3 {
+                x: 1.,
+                y: world_pos.y,
+                z: 1.,
+            }),
+            ..default()
+        };
+
         TerrainBundle {
             terrain_type,
             tile_pos,
             raycast_mesh: RaycastMesh::<Terrain>::default(),
             zoning: Zoning::None,
+            pbr_bundle,
         }
     }
 }
